page(), page_meta() and sitemap_page_walk() changed
Xysti::sitemap_page_walk() no longer returns false if it cannot find a
sitemap entry, instead it returns $page['not_found'], page() then passes
this to page_meta() which interprets this to mean the page does not
exist. The reasoning for this is the slug can now be passed between the
function enabling page_meta() to guess at a page's title by the slug
providing better <title> and breadcrumbs on error pages.